# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Checkout Store API",
    "version": "0.2.0",
    "description": "Single Serverless Framework HTTP API. Import into Apidog. Update on every endpoint change."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "serverless offline / deployed HttpApi"
    }
  ],
  "paths": {
...
    "/transactions/{id}": {
      "get": {
        "tags": ["transactions"],
        "summary": "Get transaction by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Transaction" },
          "404": { "description": "Not found" }
        }
      }
    },
    "/transactions/{id}/pay": {
      "post": {
        "tags": ["transactions"],
        "summary": "Pay PENDING transaction (payment port)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Payment outcome (APPROVED/DECLINED/ERROR)" },
          "400": { "description": "Validation" },
          "404": { "description": "Not found" },
          "422": { "description": "Invalid state" }
        }
      }
    }
  }
}